pkg/proc/internal/ebpf: Fix size of ebpf type for fn_addr
Must have been an issue overlooked in the initial implementation but
we should be using a 64-bit wide type to store the function address.

+++ pkg/proc/internal/ebpf/bpf/include/function_vals.bpf.h
@@ -32,7 +32,7 @@ typedef struct function_parameter_list {
       long long g_addr_offset;  // Offset of the Goroutine struct from the TLS segment.
       int goroutine_id;
 
-      unsigned int fn_addr;
+      unsigned long long int fn_addr;
       bool is_ret;
 
       unsigned int n_parameters;          // number of parameters.
